fix: write_to_file new example: dynamic app new example: story generator
