+++ .gitignore
@@ -33,6 +33,7 @@ news/
 .chroma/
 hierarchical_memory/
 .repomix-output.txt
+output/
 
 # PyInstaller
 #  Usually these files are written by a python script from a template
