+++ examples/03_apps/story_gen/cursor.py
